requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
avro-python3==1.10.2
azure-core==1.23.1
azure-eventhub==5.7.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-storage-blob==12.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
configparser==5.2.0
cryptography==36.0.2
fastapi==0.75.1
FastAPI-SQLAlchemy==0.2.1
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
isodate==0.6.1
msrest==0.6.21
multidict==6.0.2
nest-asyncio==1.5.5
oauthlib==3.2.0
pipfile==0.0.2
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
toml==0.10.2
tornado==4.5.3
tweepy==4.8.0
typing_extensions==4.1.1
uamqp==1.5.3
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
yarl==1.7.2
